requirements.txt
colorlover==0.3.0
einops==0.6.1
gradio==3.41.0
imageio==2.31.2
img2dataset==1.42.0
kornia==0.6.9
lpips==0.1.4
matplotlib==3.7.2
numpy==1.25.1
omegaconf==2.3.0
open-clip-torch==2.20.0
opencv-python==4.6.0.66
Pillow==9.5.0
pytorch-fid==0.3.0
pytorch-lightning==2.0.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
socksio==1.0.0
tensorboard==2.14.0
timm==0.9.2
tokenizers==0.13.3
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
transformers==4.30.2
xformers==0.0.22.post7

